fix: product

+++ src/repository/category.repository.ts
@@ -7,16 +7,26 @@ export const categoryRepository = {
     create: async (data: Prisma.CategoryCreateInput) => {
         return prisma.category.create({
             data,
+            include: {
+                images: true,
+            },
         });
     },
 
     findAll: async () => {
-        return prisma.category.findMany();
+        return prisma.category.findMany({
+            include: {
+                images: true,
+            },
+        });
     },
 
     findById: async (id: string) => {
         return prisma.category.findUnique({
             where: { id },
+            include: {
+                images: true,
+            },
         });
     },
 
@@ -30,12 +40,39 @@ export const categoryRepository = {
         return prisma.category.update({
             where: { id },
             data,
+            include: {
+                images: true,
+            },
         });
     },
 
     delete: async (id: string) => {
+        const category = await prisma.category.findUnique({
+            where: {
+                id,
+            },
+            include: {
+                products: true,
+            },
+        });
+
+
+        if (!category) {
+            throw new Error("Category not found");
+        }
+
+
+        if (category.products.length > 0) {
+            throw new Error(
+                "Cannot delete category because it has products"
+            );
+        }
+
+
         return prisma.category.delete({
-            where: { id },
+            where: {
+                id,
+            },
         });
     },
 }
